[MISC] Split type list and type pack traits.
Moves the traits for type paks into its own submodule.
Adapts all headers, snippets and tests to reflect the structural change.
