+++ include/seqan3/alphabet/composite/alphabet_tuple_base.hpp
@@ -26,6 +26,9 @@
 #include <seqan3/core/concept/tuple.hpp>
 #include <seqan3/core/type_traits/transformation_trait_or.hpp>
 #include <seqan3/utility/detail/integer_traits.hpp>
+#include <seqan3/utility/type_list/traits.hpp>
+#include <seqan3/utility/type_pack/traits.hpp>
+#include <seqan3/std/concepts>
 
 namespace seqan3::detail
 {
