+++ include/seqan3/core/algorithm/configuration.hpp
@@ -19,9 +19,9 @@
 #include <seqan3/core/algorithm/concept.hpp>
 #include <seqan3/core/algorithm/configuration_utility.hpp>
 #include <seqan3/core/algorithm/pipeable_config_element.hpp>
-#include <seqan3/utility/type_list/traits.hpp>
 #include <seqan3/core/tuple_utility.hpp>
 #include <seqan3/utility/type_list/type_list.hpp>
+#include <seqan3/utility/type_pack/traits.hpp>
 #include <seqan3/std/concepts>
 
 namespace seqan3
