+++ include/seqan3/core/pod_tuple.hpp
@@ -16,7 +16,7 @@
 #include <type_traits>
 
 #include <seqan3/core/platform.hpp>
-#include <seqan3/utility/type_list/traits.hpp>
+#include <seqan3/utility/type_pack/traits.hpp>
 
 namespace seqan3
 {
