+++ test/snippet/utility/type_pack/pack_traits_replace_at.cpp
@@ -1,4 +1,4 @@
-#include <seqan3/utility/type_list/traits.hpp>
+#include <seqan3/utility/type_pack/traits.hpp>
 
 int main()
 {
