+++ test/unit/utility/type_list/CMakeLists.txt
@@ -1,1 +1,2 @@
 seqan3_test (type_list_test.cpp)
+seqan3_test (type_list_traits_test.cpp)
