+++ test/unit/utility/type_pack/CMakeLists.txt
@@ -1,0 +1,1 @@
+seqan3_test (type_pack_traits_test.cpp)
